% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_anagram.R
\name{is_anagram}
\alias{is_anagram}
\title{Check for Anagrams}
\usage{
is_anagram(string, terms, value = FALSE, any_len = FALSE,
  ignore_spaces = FALSE, ignore_case = FALSE)
}
\arguments{
\item{string}{Char string.}

\item{terms}{Character vector.}

\item{value}{Logical, return anagram matches (TRUE), or logical vector
(FALSE). Default value is FALSE.}

\item{any_len}{Logical, look for anagrams that are any length (TRUE), or
restrict to same length anagrams (FALSE). Default value is FALSE.}

\item{ignore_spaces}{logical, ignores spaces. Default value is FALSE.}

\item{ignore_case}{logical, ignores case. Default value is FALSE.}
}
\value{
If \code{value} is FALSE, return logical vector with length equal
 to that of \code{terms}. Otherwise, return character vector of anagram matches.
}
\description{
Given an input string and a character vector, check if string is an anagram
to each element of the character vector.
}
\examples{
str_to_check <- c("cats are great", "tacs", "frogs", "cats", "ts")
is_anagram("cats", str_to_check)
is_anagram("cats", str_to_check, value = TRUE)
is_anagram("cats", str_to_check, any_len = TRUE)
is_anagram("c a t s", str_to_check, ignore_spaces = TRUE)
is_anagram("CATs", str_to_check, ignore_case = TRUE)

}
