% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_anagram.R
\name{is_anagram}
\alias{is_anagram}
\title{Check for Anagrams}
\usage{
is_anagram(string, terms, any_len = FALSE, ignore_spaces = FALSE,
  ignore_case = FALSE)
}
\arguments{
\item{string}{Char string.}

\item{terms}{Character vector.}

\item{any_len}{Logical, should the function be looking for anagrams that
are any length, or be restricted to anagrams that are the same length as
input arg \code{string}. Default value is FALSE.}

\item{ignore_spaces}{logical, should spaces be ignored when searching for
anagrams. Default value is FALSE.}

\item{ignore_case}{logical, should case (lower/upper) be ignored when
searching for anagrams. Default value is FALSE.}
}
\value{
Logical vector, with length equal to that of \code{terms}.
}
\description{
Given an input string and a character vector, for each element of the
character vector, check if input string is an anagram of the given element.
}
\examples{
str_to_check <- c("cats are great", "tacs", "frogs", "cats", "ts")
is_anagram("cats", str_to_check)
is_anagram("cats", str_to_check, any_len = TRUE)
is_anagram("c a t s", str_to_check, ignore_spaces = TRUE)
is_anagram("CATs", str_to_check, ignore_case = TRUE)

}
